update to problem 207

+++ leetcode_from_problem107/0204-Count Primes.cpp
@@ -16,7 +16,7 @@ public:
 	bool isPrime(int num) {
 		if (num == 2 || num == 3) return 1;
 		if (num % 6 != 1 && num % 6 != 5) return 0;
-		int tmp = sqrt(num);
+		int tmp = (int)sqrt(num);
 		for (int i = 5; i <= tmp; i += 6)
 			if (num % i == 0 || num % (i + 2) == 0) return 0;
 		return 1;
@@ -38,7 +38,7 @@ public:
 	bool isPrime(int num) {
 		if (num == 2 || num == 3) return 1;
 		if (num % 6 != 1 && num % 6 != 5) return 0;
-		int tmp = sqrt(num);
+		int tmp = (int)sqrt(num);
 		for (int i = 5; i <= tmp; i += 6)
 			if (num % i == 0 || num % (i + 2) == 0) return 0;
 		return 1;
